Pixbuf & Co. depend on gdk_pixbuf_* build tag now
